Replace placeholder models with the real thing
Also did a fair bit of refactoring to how we handle collision detection, so
that now each object (including laser beams and bombs) is responsible for what
it does when it collides with something. This is a lot easier to keep track of
than before.

Also, now you can shoot down enemy bombs thanks to the change in collision
handling logic. This fixes issue #6. I think the change has also fixed issue #4
as the conning tower is now treated the same as any other part of the model.